feat(ftp): Sftp client for Secure File Transfer Protocol
We have the need to use this protocol for a SFTP server we have, and the current protocol solutions that the FTP Connector provide (FTP, FTPS) cannot cover that need.

This feature provides this capacity using JSch (Java Secure Channel)
